Add a simple exclusionary lookup/cache thing
Avoids hitting the database for any path that is outside of the longest common prefix for each drive letter. This effectively eliminates all work done by watched.dll for any paths that can be excluded by this method. It wasn't particularly heavy before, but this should be a decent benefit for very little cost (1 path stored in memory per drive letter, essentially).

In practice, when profiling explorer.exe and copying a large directory of files (which triggers a `IsMemberOf` call) outside the longest common prefix for that drive:

- Before: ~0.70% of samples were within watched.dll
- After: ~0.04% of samples were within watched.dll